+++ src/main/kotlin/com/pluxity/weekly/chat/service/ChatReadHandler.kt
@@ -13,6 +13,7 @@ import com.pluxity.weekly.epic.service.EpicService
 import com.pluxity.weekly.project.entity.ProjectStatus
 import com.pluxity.weekly.project.service.ProjectService
 import com.pluxity.weekly.task.entity.TaskStatus
+import com.pluxity.weekly.task.service.TaskReviewService
 import com.pluxity.weekly.task.service.TaskService
 import com.pluxity.weekly.team.service.TeamService
 import org.springframework.stereotype.Component
@@ -20,6 +21,7 @@ import org.springframework.stereotype.Component
 @Component
 class ChatReadHandler(
     private val taskService: TaskService,
+    private val taskReviewService: TaskReviewService,
     private val projectService: ProjectService,
     private val epicService: EpicService,
     private val teamService: TeamService,
@@ -46,7 +48,7 @@ class ChatReadHandler(
                 )
             ChatTarget.REVIEW ->
                 ChatReadResponse(
-                    pendingReviews = taskService.findPendingReviews(),
+                    pendingReviews = taskReviewService.findPendingReviews(),
                 )
         }
     }
